refactor: merge space related folder loaders into one
