+++ packages/web-app-files/src/helpers/resources.ts
@@ -16,8 +16,7 @@ import {
 } from 'web-client/src/helpers/share'
 import { extractExtensionFromFile, extractStorageId } from './resource'
 import { buildWebDavSpacesPath, extractDomSelector } from 'web-client/src/helpers/resource'
-import { SHARE_JAIL_ID } from '../services/folder'
-import { Resource, SpaceResource } from 'web-client/src/helpers'
+import { Resource, SpaceResource, SHARE_JAIL_ID } from 'web-client/src/helpers'
 import { urlJoin } from 'web-pkg/src/utils'
 
 export function renameResource(space: SpaceResource, resource: Resource, newPath: string) {
