+++ packages/web-app-files/src/services/folder/index.ts
@@ -1,9 +1,7 @@
-export * from './legacy/loaderPersonal'
-export * from './spaces/loaderSpaceGeneric'
-export * from './spaces/loaderShare'
+export * from './loaderSpace'
 export * from './loaderFavorites'
-export * from './loaderPublicFiles'
 export * from './loaderSharedViaLink'
 export * from './loaderSharedWithMe'
 export * from './loaderSharedWithOthers'
 export * from './loaderTrashbin'
+export * from './types'
