+++ packages/web-app-files/src/services/folder/types.ts
@@ -1,0 +1,3 @@
+export interface FolderLoaderOptions {
+  shareIndicators?: boolean
+}
